+++ prisma/migrations/20240321004552_change_refresh_token_to_scalar_array/migration.sql
@@ -1,9 +1,0 @@
-/*
-  Warnings:
-
-  - The `refreshToken` column on the `User` table would be dropped and recreated. This will lead to data loss if there is data in the column.
-
-*/
--- AlterTable
-ALTER TABLE "User" DROP COLUMN "refreshToken",
-ADD COLUMN     "refreshToken" TEXT[];