+++ src/HealthCheck/health-check.controller.ts
@@ -1,7 +1,9 @@
 import { Controller, Get } from "@nestjs/common";
 import { Public } from "../global/decorators/public.decorator";
+import { ApiTags } from "@nestjs/swagger";
 
 @Public()
+@ApiTags("Healthchecks")
 @Controller("health-check")
 export class HealthCheckController {
     @Get("/")