+++ src/ability/prisma-generated-types.ts
@@ -1,6 +1,13 @@
 import { Subjects } from "@casl/prisma";
 
-import { Form, ProjectIdea, User, Voyage, VoyageTeam } from "@prisma/client";
+import {
+    Form,
+    ProjectIdea,
+    User,
+    Voyage,
+    VoyageTeam,
+    TeamTechStackItem,
+} from "@prisma/client";
 
 export type PrismaSubjects = Subjects<{
     Voyage: Voyage;
@@ -8,4 +15,5 @@ export type PrismaSubjects = Subjects<{
     Form: Form;
     VoyageTeam: VoyageTeam;
     Ideation: ProjectIdea;
+    TeamTechStackItem: TeamTechStackItem;
 }>;