+++ src/ideations/ideations.controller.ts
@@ -55,7 +55,7 @@ export class IdeationsController {
         type: IdeationResponse,
     })
     @CheckAbilities({ action: Action.Create, subject: "Ideation" })
-    @Post()
+    @Post("teams/:teamId/ideations")
     createIdeation(
         @Request() req: CustomRequest,
         @Param("teamId", ParseIntPipe) teamId: number,
@@ -94,17 +94,12 @@ export class IdeationsController {
         type: IdeationVoteResponse,
     })
     @CheckAbilities({ action: Action.Create, subject: "Ideation" })
-    @Post("/:ideationId/ideation-votes")
+    @Post("ideations/:ideationId/ideation-votes")
     createIdeationVote(
         @Request() req: CustomRequest,
-        @Param("teamId", ParseIntPipe) teamId: number,
         @Param("ideationId", ParseIntPipe) ideationId: number,
     ) {
-        return this.ideationsService.createIdeationVote(
-            req,
-            teamId,
-            ideationId,
-        );
+        return this.ideationsService.createIdeationVote(req, ideationId);
     }
 
     @ApiOperation({
@@ -122,7 +117,7 @@ export class IdeationsController {
         type: TeamIdeationsResponse,
     })
     @CheckAbilities({ action: Action.Read, subject: "Ideation" })
-    @Get()
+    @Get("teams/:teamId/ideations")
     getIdeationsByVoyageTeam(
         @Param("teamId", ParseIntPipe) teamId: number,
         @Request() req: CustomRequest,
@@ -151,17 +146,15 @@ export class IdeationsController {
         type: IdeationResponse,
     })
     @CheckAbilities({ action: Action.Update, subject: "Ideation" })
-    @Patch("/:ideationId")
+    @Patch("ideations/:ideationId")
     updateIdeation(
         @Request() req: CustomRequest,
         @Param("ideationId", ParseIntPipe) ideationId: number,
-        @Param("teamId", ParseIntPipe) teamId: number,
         @Body() updateIdeationDto: UpdateIdeationDto,
     ) {
         return this.ideationsService.updateIdeation(
             req,
             ideationId,
-            teamId,
             updateIdeationDto,
         );
     }
@@ -197,13 +190,12 @@ export class IdeationsController {
         type: ConflictErrorResponse,
     })
     @CheckAbilities({ action: Action.Delete, subject: "Ideation" })
-    @Delete("/:ideationId")
+    @Delete("ideations/:ideationId")
     deleteIdeation(
         @Request() req: CustomRequest,
-        @Param("teamId", ParseIntPipe) teamId: number,
         @Param("ideationId", ParseIntPipe) ideationId: number,
     ) {
-        return this.ideationsService.deleteIdeation(req, teamId, ideationId);
+        return this.ideationsService.deleteIdeation(req, ideationId);
     }
 
     @ApiOperation({
@@ -227,17 +219,12 @@ export class IdeationsController {
         type: IdeationVoteResponse,
     })
     @CheckAbilities({ action: Action.Delete, subject: "Ideation" })
-    @Delete("/:ideationId/ideation-votes")
+    @Delete("ideations/:ideationId/ideation-votes")
     deleteIdeationVote(
         @Request() req: CustomRequest,
-        @Param("teamId", ParseIntPipe) teamId: number,
         @Param("ideationId", ParseIntPipe) ideationId: number,
     ) {
-        return this.ideationsService.deleteIdeationVote(
-            req,
-            teamId,
-            ideationId,
-        );
+        return this.ideationsService.deleteIdeationVote(req, ideationId);
     }
 
     @ApiOperation({
@@ -278,7 +265,7 @@ export class IdeationsController {
         example: 1,
     })
     @CheckAbilities({ action: Action.Manage, subject: "Ideation" })
-    @Post("/:ideationId/select")
+    @Post("teams/:teamId/ideations/:ideationId/select")
     setIdeationSelection(
         @Request() req: CustomRequest,
         @Param("teamId", ParseIntPipe) teamId: number,
@@ -318,7 +305,7 @@ export class IdeationsController {
         example: 1,
     })
     @CheckAbilities({ action: Action.Manage, subject: "all" })
-    @Post("/reset-selection")
+    @Post("teams/:teamId/ideations/reset-selection")
     resetIdeationSelection(
         @Request() req: CustomRequest,
         @Param("teamId", ParseIntPipe) teamId: number,