+++ test/techs.e2e-spec.ts
@@ -77,6 +77,7 @@ describe("Techs Controller (e2e)", () => {
                             {
                                 id: expect.any(Number),
                                 name: expect.any(String),
+                                isSelected: expect.any(Boolean),
                                 teamTechStackItemVotes: expect.any(Array),
                             },
                         ]),
@@ -102,6 +103,38 @@ describe("Techs Controller (e2e)", () => {
                         ]),
                     );
                 });
+        });
+        it("should return 404 if invalid teamId provided", async () => {
+            const teamId: number = 9999999;
+
+            return request(app.getHttpServer())
+                .get(`/voyages/teams/${teamId}/techs`)
+                .set("Cookie", accessToken)
+                .expect(404)
+                .expect("Content-Type", /json/);
+        });
+        it("should return 401 unauthorized if not logged in", async () => {
+            const teamId: number = 2;
+
+            return request(app.getHttpServer())
+                .get(`/voyages/teams/${teamId}/techs`)
+                .set("Authorization", `Bearer ${undefined}`)
+                .expect(401)
+                .expect("Content-Type", /json/);
+        });
+        it("should return 403 if a user of other team tries to access tech stack items", async () => {
+            const { access_token } = await loginAndGetTokens(
+                "dan@random.com",
+                "password",
+                app,
+            );
+            const teamId: number = 2;
+
+            return request(app.getHttpServer())
+                .get(`/voyages/teams/${teamId}/techs`)
+                .set("Cookie", access_token)
+                .expect(403)
+                .expect("Content-Type", /json/);
         });
     });
     describe("POST voyages/teams/:teamId/techs - add new tech item", () => {
@@ -144,7 +177,7 @@ describe("Techs Controller (e2e)", () => {
 
         it("should return 400 if invalid team member id provided", async () => {
             const teamId: number = 2;
-            const teamMemberId: number = 4;
+            const teamMemberId: number = 5;
 
             return request(app.getHttpServer())
                 .post(`/voyages/teams/${teamId}/techs`)
@@ -179,6 +212,25 @@ describe("Techs Controller (e2e)", () => {
                         }),
                     );
                 });
+        });
+        it("should return 403 if a user of other team tries to add a tech stack item", async () => {
+            const { access_token } = await loginAndGetTokens(
+                "dan@random.com",
+                "password",
+                app,
+            );
+            const teamId: number = 2;
+            const teamMemberId: number = 9;
+
+            return request(app.getHttpServer())
+                .post(`/voyages/teams/${teamId}/techs`)
+                .set("Cookie", access_token)
+                .send({
+                    techName: newTechName,
+                    techCategoryId: 1,
+                    voyageTeamMemberId: teamMemberId,
+                })
+                .expect(403);
         });
 
         it("should return 404 if invalid teamId provided", async () => {
@@ -312,6 +364,22 @@ describe("Techs Controller (e2e)", () => {
                 })
                 .expect(403);
         });
+        it("should return 403 if a user of other team tries to update a tech stack item", async () => {
+            const { access_token } = await loginAndGetTokens(
+                "dan@random.com",
+                "password",
+                app,
+            );
+            const techId: number = 1;
+
+            return request(app.getHttpServer())
+                .patch(`/voyages/techs/${techId}`)
+                .set("Cookie", access_token)
+                .send({
+                    techName: updatedTechName,
+                })
+                .expect(403);
+        });
         it("should return 401 unauthorized if not logged in", async () => {
             const techId: number = 5;
 
@@ -375,6 +443,19 @@ describe("Techs Controller (e2e)", () => {
                     );
                 });
         });
+        it("should return 403 if a user of other team tries to delete a tech stack item", async () => {
+            const { access_token } = await loginAndGetTokens(
+                "dan@random.com",
+                "password",
+                app,
+            );
+            const techId: number = 1;
+
+            return request(app.getHttpServer())
+                .delete(`/voyages/techs/${techId}`)
+                .set("Cookie", access_token)
+                .expect(403);
+        });
         it("should return 401 if user is not logged in", async () => {
             const techId: number = 5;
 
@@ -433,6 +514,30 @@ describe("Techs Controller (e2e)", () => {
                         }),
                     );
                 });
+        });
+
+        it("should return 404 if tech item to vote for does not exist", async () => {
+            const techId: number = 9999999;
+
+            return request(app.getHttpServer())
+                .post(`/voyages/techs/${techId}/vote`)
+                .set("Cookie", accessToken)
+                .expect(404)
+                .expect("Content-Type", /json/);
+        });
+        it("should return 403 if a user of other team tries to vote for a tech stack item", async () => {
+            const { access_token } = await loginAndGetTokens(
+                "dan@random.com",
+                "password",
+                app,
+            );
+            const techId: number = 1;
+
+            return request(app.getHttpServer())
+                .post(`/voyages/techs/${techId}/vote`)
+                .set("Cookie", access_token)
+                .expect(403)
+                .expect("Content-Type", /json/);
         });
 
         it("should return 409 if user vote for tech already exists", async () => {
@@ -528,6 +633,29 @@ describe("Techs Controller (e2e)", () => {
                     );
                 });
         });
+        it("should return 404 if tech item to vote for does not exist", async () => {
+            const techId: number = 9999999;
+
+            return request(app.getHttpServer())
+                .delete(`/voyages/techs/${techId}/vote`)
+                .set("Cookie", accessToken)
+                .expect(404)
+                .expect("Content-Type", /json/);
+        });
+        it("should return 403 if a user of other team tries to vote for a tech stack item", async () => {
+            const { access_token } = await loginAndGetTokens(
+                "dan@random.com",
+                "password",
+                app,
+            );
+            const techId: number = 1;
+
+            return request(app.getHttpServer())
+                .delete(`/voyages/techs/${techId}/vote`)
+                .set("Cookie", access_token)
+                .expect(403)
+                .expect("Content-Type", /json/);
+        });
 
         it("should return 404 if vote to delete does not exist", async () => {
             const techId: number = 6;
@@ -618,8 +746,8 @@ describe("Techs Controller (e2e)", () => {
                 .expect(400);
         });
 
-        it("should return 400 invalid team id provided", async () => {
-            const teamId: number = 3;
+        it("should return 404 invalid team id provided", async () => {
+            const teamId: number = 999999;
 
             return request(app.getHttpServer())
                 .patch(`/voyages/teams/${teamId}/techs/selections`)
@@ -637,7 +765,34 @@ describe("Techs Controller (e2e)", () => {
                         },
                     ],
                 })
-                .expect(400);
+                .expect(404);
+        });
+        it("should return 403 if a user does not belong to the same team", async () => {
+            const { access_token } = await loginAndGetTokens(
+                "dan@random.com",
+                "password",
+                app,
+            );
+
+            const teamId: number = 2;
+
+            return request(app.getHttpServer())
+                .patch(`/voyages/teams/${teamId}/techs/selections`)
+                .set("Cookie", access_token)
+                .send({
+                    categories: [
+                        {
+                            categoryId: 1,
+                            techs: [
+                                {
+                                    techId: 1,
+                                    isSelected: true,
+                                },
+                            ],
+                        },
+                    ],
+                })
+                .expect(403);
         });
 
         it("should return 401 unauthorized if not logged in", async () => {